Started changing copyright and credits
Stumbled upon the mass of GOOGLE_ constants used by the build system and decided it was as good a moment as any to start changing this. Still looking for the bundle ID and other things I would rather change to make sure my fork does not collide with GQSB when run concurrently.